# Conflicts:
#	src/SubscriptionBundle/Controller/Actions/ConsentPageSubscribeAction.php

+++ src/App/Domain/Service/CrossSubscriptionAPI/ApiConnector.php
@@ -27,6 +27,7 @@ class ApiConnector
 
     /**
      * SubscribeExternalAPICheck constructor.
+     *
      * @param Client $guzzleClient
      */
     public function __construct(Client $guzzleClient, string $apiLink)
@@ -37,23 +38,30 @@ class ApiConnector
 
     public function checkIfExists(string $msisdn, int $carrierId): bool
     {
-
-        $result   = $this->guzzleClient->get(sprintf('%s/msisdn/%s/%s', $this->apiLink, $carrierId, $msisdn));
-        $response = json_decode($result->getBody(), true);
-        $isExists = $response['isExist'] ?? false;
-        return (bool)$isExists;
+        if(filter_var($this->apiLink, FILTER_VALIDATE_URL)) {
+            $result   = $this->guzzleClient->get(sprintf('%s/msisdn/%s/%s', $this->apiLink, $carrierId, $msisdn));
+            $response = json_decode($result->getBody(), true);
+            $isExists = $response['isExist'] ?? false;
+            return (bool)$isExists;
+        }
+        return false;
     }
 
     public function registerSubscription(string $msisdn, int $carrierId): void
     {
-        $this->guzzleClient->post(sprintf('%s/msisdn', $this->apiLink), $options = [
-            RequestOptions::JSON => ['carrierId' => $carrierId, 'msisdn' => $msisdn],
-        ]);
+        if(filter_var($this->apiLink, FILTER_VALIDATE_URL)) {
+            $this->guzzleClient->post(sprintf('%s/msisdn', $this->apiLink), $options = [
+                RequestOptions::JSON => ['carrierId' => $carrierId, 'msisdn' => $msisdn],
+            ]);
+        }
     }
+
     public function deregisterSubscription(string $msisdn, int $carrierId): void
     {
-        $this->guzzleClient->delete(sprintf('%s/msisdn', $this->apiLink), $options = [
-            RequestOptions::JSON => ['carrierId' => $carrierId, 'msisdn' => $msisdn],
-        ]);
+        if(filter_var($this->apiLink, FILTER_VALIDATE_URL)) {
+            $this->guzzleClient->delete(sprintf('%s/msisdn', $this->apiLink), $options = [
+                RequestOptions::JSON => ['carrierId' => $carrierId, 'msisdn' => $msisdn],
+            ]);
+        }
     }
 }
